pyproject.toml
[project]
name = "intikhab"
version = "0.1.0"
description = "Electronic Voting System - Secure, Transparent, Democratic"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=45.0.7",
    "django>=5.2.6",
    "django-allauth>=65.11.2",
    "django-environ>=0.12.0",
    "lightphe>=0.0.15",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "sendgrid>=6.12.4",
    "gunicorn>=21.2.0",
    "whitenoise>=6.6.0",
    "psycopg2-binary>=2.9.9",
    "dj-database-url>=2.1.0",
    "django-htmx>=1.24.1",
    "pillow>=11.3.0",
    "factory-boy>=3.3.3",
]
